# lib/api-spec/openapi.yml
openapi: 3.1.0
info:
  title: TestCaseHub API
  version: "1.0.0"
  description: UAT Test Case Management API
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [ADMIN, USER] }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...
    Attachment:
      type: object
      properties:
        id: { type: integer }
        entity_type: { type: string }
        entity_id: { type: integer }
        field: { type: string, nullable: true }
        file_name: { type: string }
        file_url: { type: string }
        file_type: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    TestRunChecklistItem:
      type: object
      properties:
        id: { type: integer }
        test_run_id: { type: integer }
        item_text: { type: string }
        is_checked: { type: boolean }
        checked_by_user_id: { type: integer, nullable: true }
        checked_at: { type: string, format: date-time, nullable: true }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
...
    Bug:
      type: object
      properties:
        id: { type: integer }
        project_id: { type: integer }
        defect_id: { type: integer }
        bug_number: { type: integer }
        support_ticket_number: { type: string, nullable: true }
        assigned_developer_id: { type: integer, nullable: true }
        status: { type: string, enum: [OPEN, ASSIGNED, IN_PROGRESS, RESOLVED, FAILED_TO_RESOLVE, CLOSED, REOPENED] }
        developer_notes: { type: string, nullable: true }
        failed_to_resolve_reason: { type: string, nullable: true }
        root_cause_category: { type: string, nullable: true }
        opened_at: { type: string, format: date-time }
        assigned_at: { type: string, format: date-time, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        test_at: { type: string, format: date-time, nullable: true }
        failed_to_resolve_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    $ref: "#/components/schemas/User"
  /auth/register:
    post:
      summary: Register (Admin only)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                name: { type: string }
                email: { type: string }
                role: { type: string, enum: [ADMIN, USER] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /test-runs/{testRunId}/checklist:
    get:
      summary: Get test run checklist
      tags: [TestRunChecklist]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestRunChecklistItem"
    post:
      summary: Add checklist item
      tags: [TestRunChecklist]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemText: { type: string }
      responses:
        "201":
          description: Created
  /test-runs/{testRunId}/checklist/{itemId}:
    patch:
      summary: Update checklist item
      tags: [TestRunChecklist]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isChecked: { type: boolean }
      responses:
        "200":
          description: OK
    delete:
      summary: Delete checklist item
      tags: [TestRunChecklist]
      parameters:
        - in: path
          name: testRunId
          required: true
          schema:
            type: integer
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
  /projects/{projectId}/bugs:
    get:
      summary: List bugs for project
      tags: [Bugs]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bug"
  /bugs/{bugId}:
    get:
      summary: Get bug by id
      tags: [Bugs]
      parameters:
        - in: path
          name: bugId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bug"
  /bugs/{bugId}/assign:
    patch:
      summary: Assign bug to developer
      tags: [Bugs]
      parameters:
        - in: path
          name: bugId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                developerId: { type: integer }
                supportTicketNumber: { type: string }
      responses:
        "200":
          description: OK
  /bugs/{bugId}/status:
    patch:
      summary: Update bug status
      tags: [Bugs]
      parameters:
        - in: path
          name: bugId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
                reason: { type: string }
      responses:
        "200":
          description: OK
  /bugs/{bugId}/notes:
    patch:
      summary: Add / bug notes
      tags: [Bugs]
      parameters:
        - in: path
          name: bugId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
                rootCauseCategory: { type: string }
      responses:
        "200":
          description: OK
  /bugs/{bugId}/reassign:
    patch:
      summary: Reassign bug developer
      tags: [Bugs]
      parameters:
        - in: path
          name: bugId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                developerId: { type: integer }
      responses:
        "200":
          description: OK
...
  /dashboard/developer/{userId}/bugs:
    get:
      summary: Developer's bugs
      tags: [Dashboard]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /upload:
    post:
      summary: Upload file
      tags: [Uploads]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileUrl: { type: string }
                  fileName: { type: string }
                  fileType: { type: string }
  /attachments:
    post:
      summary: Create attachment record
      tags: [Uploads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Attachment"
      responses:
        "201":
          description: Created
  /attachments/{entityType}/{entityId}:
    get:
      summary: List attachments for entity
      tags: [Uploads]
      parameters:
        - in: path
          name: entityType
          required: true
          schema:
            type: string
        - in: path
          name: entityId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
  /attachments/{attachmentId}:
    delete:
      summary: Delete attachment
      tags: [Uploads]
      parameters:
        - in: path
          name: attachmentId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
...
security:
  - BearerAuth: []